Make searches for seats return flight details, including full airports, in nested objects.
As required (for the user-facing APIs that call this one) by the API spec.

Because the tests still pass after I make this change, without my making
any changes _to_ the tests, I am not quite certain that this is quite
right. Tests should be extended to cover the details of what a search
returns. But that's beyond my scope for today.