Make seat-details query return flight details, including full airports, in nested objects.
As with search results, this is required (for user-facing API callers) by
the API spec. And also as with search results, this commit does not change
any code path exercised by the tests, so I do not have complete confidence
that this is correct, and I note that tests should be extended (later) to
cover the details of what a seat-details query should return.